+++ CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added `PybricksHub.stdout_observable` property ([support#1038]).
+
 ### Fixed
 - Fixed endline in `PybricksHub.write_line()`.
+
+[support#1038]: https://github.com/orgs/pybricks/discussions/1038
 
 ## [1.0.0-alpha.44] - 2023-04-20